+++ plugins/types/src/entities/index.ts
@@ -5,8 +5,16 @@ export {EvaluatorData} from "./evaluator.data";
 export {Model} from "./model";
 export {Notification} from "./notification";
 export {AnalysisResult} from "./analysisResult";
-export {type ProviderResponse, type ProviderResponseView, type ProviderResponseMetadata, type ProviderResponseMetrics, ProviderResponseStatus} from "./provider.response";
-export {type ProviderRequest, type ProviderAuditRequest, type ProviderRequestMetadata, type ProviderEnvelope} from "./provider.request";
+export {
+    type ProviderResponse,
+    type ProviderResponseView,
+    type ProviderResponseMetadata,
+    type ProviderResponseMetrics,
+    ProviderResponseStatus
+} from "./provider.response";
+export {
+    type ProviderRequest, type ProviderAuditRequest, type ProviderRequestMetadata, type ProviderEnvelope
+} from "./provider.request";
 export {Application, type AccessLevel} from "./application";
 export {FinishReason} from "./finish.reason";
 export {HoloToken} from "./holo.token";